+++ json_serializable/lib/src/type_helpers/config_types.dart
@@ -61,6 +61,7 @@ class ClassConfig {
   final bool createJsonKeys;
   final bool createPerFieldToJson;
   final bool createJsonSchema;
+  final bool dateTimeUtc;
   final bool disallowUnrecognizedKeys;
   final bool explicitToJson;
   final FieldRename fieldRename;
@@ -81,6 +82,7 @@ class ClassConfig {
     required this.createJsonKeys,
     required this.createPerFieldToJson,
     required this.createJsonSchema,
+    required this.dateTimeUtc,
     required this.disallowUnrecognizedKeys,
     required this.explicitToJson,
     required this.fieldRename,
@@ -110,6 +112,7 @@ class ClassConfig {
         createFactory:
             config.createFactory ?? ClassConfig.defaults.createFactory,
         createToJson: config.createToJson ?? ClassConfig.defaults.createToJson,
+        dateTimeUtc: config.dateTimeUtc ?? ClassConfig.defaults.dateTimeUtc,
         ignoreUnannotated:
             config.ignoreUnannotated ?? ClassConfig.defaults.ignoreUnannotated,
         explicitToJson:
@@ -125,7 +128,6 @@ class ClassConfig {
         disallowUnrecognizedKeys:
             config.disallowUnrecognizedKeys ??
             ClassConfig.defaults.disallowUnrecognizedKeys,
-        // TODO typeConverters = []
       );
 
   /// An instance of [JsonSerializable] with all fields set to their default
@@ -140,6 +142,7 @@ class ClassConfig {
     createJsonKeys: false,
     createPerFieldToJson: false,
     createJsonSchema: false,
+    dateTimeUtc: false,
     disallowUnrecognizedKeys: false,
     explicitToJson: false,
     fieldRename: FieldRename.none,
@@ -166,6 +169,7 @@ class ClassConfig {
     fieldRename: fieldRename,
     enumFieldRename: enumFieldRename,
     disallowUnrecognizedKeys: disallowUnrecognizedKeys,
+    dateTimeUtc: dateTimeUtc,
     // TODO typeConverters = []
   );
 }