+++ json_serializable/test/json_serializable_test.dart
@@ -164,6 +164,7 @@ const _expectedAnnotatedTests = {
   'annotatedMethod',
   'theAnswer',
   'JsonSchemaTestClass',
+  'DateTimeUtcTestClass',
 };
 
 const _expectedSchemaTests = {