+++ json_serializable/test/test_sources/test_sources.dart
@@ -18,6 +18,7 @@ class ConfigurationImplicitDefaults {
   createFieldMap: false,
   createJsonKeys: false,
   createPerFieldToJson: false,
+  dateTimeUtc: false,
   disallowUnrecognizedKeys: false,
   enumFieldRename: FieldRename.none,
   explicitToJson: false,